fix(Spacing + Positioning): Key sizes and positon of special characters key
Moved the special characters key to the bottom left of the layout so that the caps lock key could occupy its full normal space. Tweaked the size of most of the keys so that they more closely mimic my keyboard, and to ensure that the two alternative layout keys along the bottom do not interfere with the arrow-keys positioning.